% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clean_whitespace.R
\name{clean_whitespace}
\alias{clean_whitespace}
\title{Clean-Up Whitespace in Character Vectors}
\usage{
clean_whitespace(x)
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
Clean-Up Whitespace in Character Vectors
}
\examples{
\dontrun{
clean_vec <- clean_whitespace(x = dirty_vec)
}
}
