% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{destroy_twm_ngram_db}
\alias{destroy_twm_ngram_db}
\title{Destroy 'TWM N-Gram Dictionary' PSQL Database}
\usage{
destroy_twm_ngram_db()
}
\value{
TRUE/FALSE
}
\description{
Destroy 'TWM N-Gram Dictionary' PSQL Database
}
\examples{
\dontrun{
destroy_twm_ngram_db()
status <- destroy_twm_ngram_db()
}
}
