% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-df2ngrams.R
\name{df2ngrams}
\alias{df2ngrams}
\title{Generate N-Grams for Character Columns of DataFrame}
\usage{
df2ngrams(df, n = 1000, split = " ")
}
\arguments{
\item{df}{data.frame}

\item{n}{numeric}

\item{split}{character}
}
\value{
list
}
\description{
Generate N-Grams for Character Columns of DataFrame
}
\examples{
\dontrun{
 output <- df2ngrams(df = edap_upc)
}
}
