% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{drop_tmp_tbl}
\alias{drop_tmp_tbl}
\title{Drop Temporary Table from Public Schema}
\usage{
drop_tmp_tbl(tbl)
}
\arguments{
\item{tbl}{character}
}
\description{
Drop Temporary Table from Public Schema
}
\examples{
\dontrun{
drop_tmp_tbl(tbl = 'tbl_fasdvjcaf')
}
}
