% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expect_data_type.R
\name{expect_data_type}
\alias{expect_data_type}
\title{Verify that R Object meets Data Type Constraints}
\usage{
expect_data_type(
  obj,
  type,
  not_empty,
  is_scalar,
  allow_na,
  check_names,
  throw_err
)
}
\arguments{
\item{obj}{R Object}

\item{type}{character - desired data type / 'character', 'numeric' or 'logical'}

\item{not_empty}{TRUE/FALSE - specifies if 'length(obj)' must be non-zero}

\item{is_scalar}{TRUE/FALSE - specifies if 'length(obj)' must equal 1}

\item{allow_na}{TRUE/FALSE - specifies if NA elements are allowed in 'obj'}

\item{check_names}{TRUE/FALSE - specifies if 'names(obj)' should meet same criteria}

\item{throw_err}{TRUE/FALSE - specifies if function should return or throw error}
}
\value{
R Object
}
\description{
Verify that R Object meets Data Type Constraints
}
\examples{
\dontrun{
expect_data_type(obj, type, not_empty, is_scalar, allow_na, check_names, throw_err)
}
}
