% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-process_ngram_data.R
\name{gen_ngram_data}
\alias{gen_ngram_data}
\title{Generate Temporary Table of N-Grams, Return Table Name}
\usage{
gen_ngram_data(table)
}
\arguments{
\item{table}{character}
}
\value{
character
}
\description{
Generate Temporary Table of N-Grams, Return Table Name
}
\examples{
\dontrun{
 temp_table_name <- gen_ngram_data(table = 'tbl_gfdasgfdv')
}
}
