% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-gen_ngrams.R
\name{gen_ngrams}
\alias{gen_ngrams}
\title{gen List of N-Grams from a List of Tokens (stored as character vectors)}
\usage{
gen_ngrams(tokens, n = 3, split = " ")
}
\arguments{
\item{tokens}{character}

\item{n}{integer}

\item{split}{character}
}
\value{
list
}
\description{
gen List of N-Grams from a List of Tokens (stored as character vectors)
}
\examples{
\dontrun{
 output_ngrams <- gen_ngrams(tokens = token_list, n = 3, split = ' ')
}
}
