% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-gen_tokens.R
\name{gen_tokens}
\alias{gen_tokens}
\title{Generate List of Tokens from a Character Vector and String Delimiter}
\usage{
gen_tokens(string, split = " ")
}
\arguments{
\item{string}{character}

\item{split}{character}
}
\value{
list
}
\description{
Generate List of Tokens from a Character Vector and String Delimiter
}
\examples{
\dontrun{
output_tokens <- gen_tokens(string = input_vec, split = ' ')
}
}
