% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{gen_unq_name}
\alias{gen_unq_name}
\title{Generate Random Table Name}
\usage{
gen_unq_name(use_date = TRUE)
}
\arguments{
\item{use_date}{logical}
}
\value{
character
}
\description{
Generate Random Table Name
}
\examples{
\dontrun{
output <- gen_unq_name(use_date = TRUE)
}
}
