% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-edap_test.R
\name{get_edap_item_code}
\alias{get_edap_item_code}
\title{Query Item Details from Google BigQuery 'EDAP - edw'}
\usage{
get_edap_item_code()
}
\value{
data.frame
}
\description{
Query Item Details from Google BigQuery 'EDAP - edw'
}
\examples{
\dontrun{
output <- get_edap_item_code()
}
}
