% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-edap_test.R
\name{get_edap_upc}
\alias{get_edap_upc}
\title{Query UPC-to-Item Mappings from Google BigQuery 'EDAP - edw'}
\usage{
get_edap_upc()
}
\value{
data.frame
}
\description{
Query UPC-to-Item Mappings from Google BigQuery 'EDAP - edw'
}
\examples{
\dontrun{
output <- get_edap_upc()
}
}
