% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-edap_test.R
\name{get_prod_upc}
\alias{get_prod_upc}
\title{Get Test UPC Data from Excel}
\usage{
get_prod_upc()
}
\value{
data.frame
}
\description{
Get Test UPC Data from Excel
}
\examples{
\dontrun{
output <- get_prod_upc()
}
}
