% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gsql_db_connect.R
\name{gsql_db_connect}
\alias{gsql_db_connect}
\title{Connect to Google BigQuery Database}
\usage{
gsql_db_connect(db)
}
\arguments{
\item{db}{character - Name of an existing database}
}
\value{
DBIConnection - R Object handle for the newly opened connection
}
\description{
Connect to Google BigQuery Database
}
\examples{
\dontrun{
conn <- gsql_db_connect('edw')
}
}
