% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{init_twm_ngram_db}
\alias{init_twm_ngram_db}
\title{Initialize 'TWM N-Gram Dictionary' PSQL Database}
\usage{
init_twm_ngram_db()
}
\value{
TRUE/FALSE
}
\description{
Initialize 'TWM N-Gram Dictionary' PSQL Database
}
\examples{
\dontrun{
init_twm_ngram_db()
status <- init_twm_ngram_db()
}
}
