% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{init_twm_ngram_func}
\alias{init_twm_ngram_func}
\title{Initialize 'TWM N-Gram Dictionary' PSQL Functions}
\usage{
init_twm_ngram_func()
}
\value{
logical
}
\description{
Initialize 'TWM N-Gram Dictionary' PSQL Functions
}
\examples{
\dontrun{
init_twm_ngram_func()
status <- init_twm_ngram_func()
}
}
