% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{init_twm_ngram_indexes}
\alias{init_twm_ngram_indexes}
\title{Initialize 'TWM N-Gram Dictionary' PSQL Indexes}
\usage{
init_twm_ngram_indexes(schema)
}
\arguments{
\item{schema}{character}
}
\value{
logical
}
\description{
Initialize 'TWM N-Gram Dictionary' PSQL Indexes
}
\examples{
\dontrun{
init_twm_ngram_indexes()
status <- init_twm_ngram_indexes()
}
}
