% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{init_twm_ngram_proc}
\alias{init_twm_ngram_proc}
\title{Initialize 'TWM N-Gram Dictionary' PSQL Procedures}
\usage{
init_twm_ngram_proc()
}
\value{
logical
}
\description{
Initialize 'TWM N-Gram Dictionary' PSQL Procedures
}
\examples{
\dontrun{
init_twm_ngram_proc()
status <- init_twm_ngram_proc()
}
}
