% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{init_twm_ngram_schema}
\alias{init_twm_ngram_schema}
\title{Initialize 'TWM N-Gram Dictionary' PSQL Schemas}
\usage{
init_twm_ngram_schema()
}
\value{
TRUE/FALSE
}
\description{
Initialize 'TWM N-Gram Dictionary' PSQL Schemas
}
\examples{
\dontrun{
init_twm_ngram_schema()
status <- init_twm_ngram_schema()
}
}
