% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{init_twm_ngram_tables}
\alias{init_twm_ngram_tables}
\title{Initialize 'TWM N-Gram Dictionary' PSQL Tables}
\usage{
init_twm_ngram_tables(schema)
}
\arguments{
\item{schema}{character}
}
\value{
logical
}
\description{
Initialize 'TWM N-Gram Dictionary' PSQL Tables
}
\examples{
\dontrun{
init_twm_ngram_tables()
status <- init_twm_ngram_tables()
}
}
