% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-proc_data_ngram.R
\name{insert_ngram_column}
\alias{insert_ngram_column}
\title{Upsert Data into Column Table}
\usage{
insert_ngram_column(dataset, column_names, schema)
}
\arguments{
\item{dataset}{data.frame}

\item{column_names}{character}

\item{schema}{character}
}
\value{
data.frame
}
\description{
Upsert Data into Column Table
}
\examples{
\dontrun{
column <- insert_ngram_column(dataset, column_names)
}
}
