% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-proc_data_ngram.R
\name{insert_ngram_dataset}
\alias{insert_ngram_dataset}
\title{Upsert Data into Dataset Table}
\usage{
insert_ngram_dataset(data, schema)
}
\arguments{
\item{data}{data.frame}

\item{schema}{character}
}
\value{
data.frame
}
\description{
Upsert Data into Dataset Table
}
\examples{
\dontrun{
dataset <- insert_ngram_dataset(data = edap_upc, schema = 'document')
}
}
