% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-process_ngram_data.R
\name{insert_ngram_record}
\alias{insert_ngram_record}
\title{Upsert Data into Record Table}
\usage{
insert_ngram_record(dataset_id, row_count)
}
\arguments{
\item{dataset_id}{numeric/integer}

\item{row_count}{numeric/integer}
}
\description{
Upsert Data into Record Table
}
\examples{
\dontrun{
insert_ngram_record(dataset_id, row_count)
}
}
