% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{kill_db_connections}
\alias{kill_db_connections}
\title{Kill All DB Connections to 'twm_ngram_dict'}
\usage{
kill_db_connections()
}
\value{
R Object
}
\description{
Kill All DB Connections to 'twm_ngram_dict'
}
\examples{
\dontrun{
kill_db_connections()
}
}
