% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list2df.R
\name{list2df}
\alias{list2df}
\title{Convert List into Formatted DataFrame}
\usage{
list2df(x, elem_type)
}
\arguments{
\item{x}{list}

\item{elem_type}{character}
}
\value{
data.frame
}
\description{
Convert List into Formatted DataFrame
}
\examples{
\dontrun{
output_dataframe <- list2df(x = input_list, elem_type = 'character')
}
}
