% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngram_corpus_tables}
\alias{ngram_corpus_tables}
\title{Base NGRAM Tables}
\format{
A data.frame of valid table names and their creation order
}
\usage{
data(ngram_corpus_tables)
}
\description{
Report ...
}
\keyword{datasets}
