% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-process_ngram_data.R
\name{ngram_data_clean}
\alias{ngram_data_clean}
\title{Clean Data Text Columns}
\usage{
ngram_data_clean(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame
}
\description{
Clean Data Text Columns
}
\examples{
\dontrun{
ngram_data_clean(data = edap_item_code)
}
}
