% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-proc_data_ngram.R
\name{proc_data}
\alias{proc_data}
\title{Process Data}
\usage{
proc_data(data, schema)
}
\arguments{
\item{data}{data.frame}

\item{schema}{character}
}
\value{
list
}
\description{
Process Data
}
\examples{
\dontrun{
proc_data(data = edap_item_code)
}
}
