% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-proc_data_ngram.R
\name{proc_ngram}
\alias{proc_ngram}
\title{Process NGRAMs}
\usage{
proc_ngram(data, schema, data_map, data_token)
}
\arguments{
\item{data}{data.frame}

\item{schema}{character}

\item{data_map}{list}

\item{data_token}{list}
}
\description{
Process NGRAMs
}
\examples{
\dontrun{
proc_ngram(data = edap_item_code)
}
}
