% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-proc_data_ngram.R
\name{process_ngram_data}
\alias{process_ngram_data}
\title{Process N-Gram Data}
\usage{
process_ngram_data(data, schema)
}
\arguments{
\item{data}{data.frame}

\item{schema}{character}
}
\description{
Process N-Gram Data
}
\examples{
\dontrun{
 process_ngram_data(data = edap_upc, schema = 'corpus')
}
}
