% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-text2ngrams.R
\name{text2ngrams}
\alias{text2ngrams}
\title{Convert Character Vector of String Values into List of NGRAMs}
\usage{
text2ngrams(string, n = 3, split = " ")
}
\arguments{
\item{string}{character}

\item{n}{integer}

\item{split}{character}
}
\value{
list
}
\description{
Convert Character Vector of String Values into List of NGRAMs
}
\examples{
\dontrun{
output_ngrams <- text2ngrams(string = input_char_vec, n = 3, split = ' ')
}
}
