% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-psql_helper.R
\name{upload_tmp_tbl}
\alias{upload_tmp_tbl}
\title{Upload R Data.Frame to PSQL Table in 'public' schema}
\usage{
upload_tmp_tbl(data)
}
\arguments{
\item{data}{data.frame}

\item{schema}{character}
}
\value{
character
}
\description{
Upload R Data.Frame to PSQL Table in 'public' schema
}
\examples{
\dontrun{
upload_tbl_name <- upload_tmp_tbl(data = raw_data)
}
}
