% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-process_ngram_data.R
\name{upsert_ngram}
\alias{upsert_ngram}
\title{Upsert N-Gram Values and Return ID Mappings}
\usage{
upsert_ngram(table, dataset)
}
\arguments{
\item{table}{character}

\item{dataset}{integer/numeric}
}
\description{
Upsert N-Gram Values and Return ID Mappings
}
\examples{
\dontrun{
 upsert_ngram(table = table, dataset = 1)
}
}
