% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-process_ngram_data.R
\name{upsert_ngram_data}
\alias{upsert_ngram_data}
\title{Upsert N-Gram Values and Return ID Mappings}
\usage{
upsert_ngram_data(table)
}
\arguments{
\item{table}{character}
}
\description{
Upsert N-Gram Values and Return ID Mappings
}
\examples{
\dontrun{
 upsert_ngram_data(table = table)
}
}
