% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate_input.R
\name{validate_input.list}
\alias{validate_input.list}
\title{S3 Method - Validate Format of List Inputs}
\usage{
\method{validate_input}{list}(
  obj,
  list_elem_type,
  list_not_empty = TRUE,
  elem_not_empty = FALSE,
  elem_is_scalar = FALSE,
  elem_allow_na = FALSE,
  elem_check_names = FALSE,
  throw_err = TRUE,
  ...
)
}
\arguments{
\item{obj}{S3 Object}

\item{list_elem_type}{character}

\item{list_not_empty}{logical - (TRUE/FALSE)}

\item{elem_not_empty}{logical - (TRUE/FALSE)}

\item{elem_is_scalar}{logical - (TRUE/FALSE)}

\item{elem_allow_na}{logical - (TRUE/FALSE)}

\item{elem_check_names}{logical - (TRUE/FALSE)}

\item{throw_err}{logical - (TRUE/FALSE)}

\item{...}{r ellipsis}
}
\value{
logical - (TRUE/FALSE)
}
\description{
S3 Method - Validate Format of List Inputs
}
\examples{
\dontrun{
is_valid <- validate_input.list(obj = input, ...)
}
}
