% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate_input.R
\name{validate_input}
\alias{validate_input}
\title{S3 Generic - Validate Format of Function Inputs}
\usage{
validate_input(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{r ellipsis}
}
\value{
logical - (TRUE/FALSE)
}
\description{
S3 Generic - Validate Format of Function Inputs
}
\examples{
\dontrun{
is_valid <- validate_input(obj = input, ...)
}
}
