% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-gen_ngrams.R
\name{gen_ngrams}
\alias{gen_ngrams}
\title{gen List of N-Grams from a Character Vector of Strings}
\usage{
gen_ngrams(string, n = 3, split = " ")
}
\arguments{
\item{string}{character}

\item{n}{integer}

\item{split}{character}
}
\value{
list
}
\description{
gen List of N-Grams from a Character Vector of Strings
}
\examples{
\dontrun{
 output_ngrams <- gen_ngrams(string = input_char_vec, n = 3, split = ' ')
}
}
