% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-process_ngram_data.R
\name{insert_ngram_dataset}
\alias{insert_ngram_dataset}
\title{Upsert Data into Dataset Table}
\usage{
insert_ngram_dataset(data, as_lib)
}
\arguments{
\item{data}{data.frame}

\item{as_lib}{logical}
}
\value{
data.frame
}
\description{
Upsert Data into Dataset Table
}
\examples{
\dontrun{
dataset <- insert_ngram_dataset(data = edap_upc)
}
}
