% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-process_ngram_data.R
\name{process_ngram_data}
\alias{process_ngram_data}
\title{Process N-Gram Data}
\usage{
process_ngram_data(data, as_lib)
}
\arguments{
\item{data}{data.frame}

\item{as_lib}{logical}
}
\value{
data.frame
}
\description{
Process N-Gram Data
}
\examples{
\dontrun{
test <- process_ngram_data(data = edap_upc)
}
}
